Add Connect to ethereum in Navbar
Here I deviate from Gregory's template
a bit; MetaMask docs urge developers to only
establish a connection with a wallet when a
user specifically requests it:

From the MetaMask docs:
"You should only initiate a connection request
in response to direct user action, such as
clicking a button. You should never initiate a
connection request on page load."

To keep in-step with this direction, I moved
the connection code to the Navbar and initiate
the connection only when the user clicks a btn.

Thoughts:
- Should the connection code be with the Navbar?
At this stage, the easiest place to put this code
is in the Navbar; that's where the button is,
that's where the account address is being displayed.
I could get more fancy with this to further isolate
the connection code, but I don't gain much by doing
that so soon.