Add stake / unstake tokens, and issue tokens
- rename withdrawTokens to unstakeTokens
for consistency.
- add new form for unstaking.
- update all balances when transactions
are mined.
- add issue-tokens script

Not yet supported: withdrawing DappTokens...
though I have an idea:

In Korea they have a thing called "key money."
You deposit a large amount of money in the name
of a landlord in exchange for renting their
apartment. The landlord collects interest
on the deposit until the lease term is over.
Once the lease is terminated, the key money
is returned to the renter.

I've decided to fork this and turn it into
a platform where users can stake money and
direct their "DappTokens" to their favorite
content providers. To do this, I will
have to make these deposits into the
Compound protocol, I believe. A portion of
the interest earned there will go to support
chosen content providers.

I'm still looking into this.